+++ driving_log_replayer/driving_log_replayer/obstacle_segmentation_analyzer.py
@@ -21,8 +21,8 @@ from typing import Tuple
 from ament_index_python.packages import get_package_share_directory
 from driving_log_replayer_analyzer.config.obstacle_segmentation import Config
 from driving_log_replayer_analyzer.config.obstacle_segmentation import load_config
-from driving_log_replayer_analyzer.data.calc import fail_3_times_in_a_row
 from driving_log_replayer_analyzer.data.jsonl_parser import JsonlParser
+from driving_log_replayer_analyzer.data.obstacle_segmentation import fail_3_times_in_a_row
 from driving_log_replayer_analyzer.plot import PlotBase
 import yaml
