+++ driving_log_replayer_analyzer/driving_log_replayer_analyzer/config/obstacle_segmentation.py
@@ -16,7 +16,7 @@ from dataclasses import dataclass
 from pathlib import Path
 from typing import Any
 
-from driving_log_replayer_analyzer.data.non_detection import FpDistance
+from driving_log_replayer_analyzer.data.obstacle_segmentation import FpDistance
 import yaml
